+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2026-02-28
+## [0.1.0] - 2026-03-10
 
 ### This is the first release of *PsigenVision Finite State Machine*.
 
@@ -33,3 +33,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Feedback and contributions are highly welcomed!
 
 ---
+
+## [0.1.1] - 2026-03-10
+
+### This is a small refactoring with the addition of a quality of life method.
+
+### Refactored
+- Removed `GetCurrentState()` method within `StateMachine` class, replacing it with the property `CurrentState`, which performs precisely the same function.
+
+### Added
+- Introduced generic method `IsCurrentState<T>()` to `StateMachine` class, which checks if the state machine's current state, `current`, is of type `T`
+
+---
+